Registering MesosStateStore in constructor instead of StoreConfig
