--- src/js/utils/StoreMixinConfig.js
@@ -1,14 +1,0 @@
-import {StoreMixin} from 'mesosphere-shared-reactjs';
-import StoreConfig from '../constants/StoreConfig';
-
-StoreMixin.store_configure(StoreConfig);
-
-function add(storeID, config) {
-  StoreConfig[storeID] = config;
-  StoreMixin.store_configure(StoreConfig);
-}
-
-module.exports = {
-  add
-};
-
